Created more reusable component & moved code
